+++ views/startchatwindow.cpp
@@ -6,7 +6,6 @@
 #include <QLineEdit>
 #include <QString>
 #include <QByteArray>
-#include <iostream>
 #include <string>
 
 StartChatWindow::StartChatWindow(QWidget *parent, MiniZoom::AppServer *appserver)
@@ -20,7 +19,7 @@ StartChatWindow::StartChatWindow(QWidget *parent, MiniZoom::AppServer *appserver
         char buffer[1024];
 
         while (true) {
-            ssize_t bytes = recv(socket_fd, buffer, sizeof(buffer), 0);
+            int bytes = recv(socket_fd, buffer, sizeof(buffer), 0);
             qDebug() << "Received " << bytes << " bytes of data";
             if (bytes <= 0) {
                 perror("Receive error");
